Finish exceptions... for now. Incomplete until subtyping comes in
Partial implementation of a language with exceptions. Definitely not
complete until we bring subtyping into the mix, because ideally we
would like exceptions to be their own custom types matched by a catch
handler or otherwise propogated upward. This we could actually already
do with our current implementation, but leaves some room for desired
work:

- No way to name exception types
- No "value-less" exception (due to the lack of a bottom type)

So, we'll push this off until we also have a language with subtyping.