Perf: optimize function calls that do not need a closure environment
Not all functions need to be treated as closures. When we have a
function to lambda lift but it does not require a closure, we can
generate a cheaper package that just contains a function pointer and the
type tag rather than allocating space for the additional (empty)
environment.

The type of both closure and raw function packages is not differentiated
because the packages are existential and should behave transparently.