Fix type inference at reference boundaries
Type inference at places where we do assignments and derefences of
references is tricky, as we actually do want to permit covariance (by
way of consistency) to the type argument of a reference there, and not
just demand strict equivalence as when we compare two references
directly. This patches fixes the issue by adding a "TArgCw" constraint
kind, which marks a constraint that should check for consistency within
a type constructor argument, regardless of what the actual constructor
is.